modificacion 1804

+++ Entity/DatosTotales_E.cs
@@ -10,7 +10,7 @@ namespace Entity
     {
         public decimal sumatotal { get; set; }
         public int cantidatotal { get; set; }    
-        public decimal maximafrecuencia { get; set; }
+        public int maximafrecuencia { get; set; }
         public int nummediana { get; set; }
         public List<Mediana_E> ltmediana = new List<Mediana_E>();
     }
